\name{regmatches<-}
\alias{regmatches<-}
\title{blah}
\usage{
  regmatches(x, m, invert = FALSE, value) <- value
}
\arguments{
  \item{x}{a character vector}

  \item{m}{an object with match data}

  \item{invert}{a logical: if TRUE, extract the non-matched
  substrings.}

  \item{value}{an object with suitable replacment values
  for the matched or non-matched substrings (see Details)}
}
\description{
  blah
}

